+++ src/joyride/extension.cljs
@@ -3,6 +3,7 @@
    ["vscode" :as vscode]
    [joyride.content-utils :as content-utils]
    [joyride.db :as db]
+   [joyride.flare.sidebar-provider :as flare-sidebar]
    [joyride.getting-started :as getting-started]
    [joyride.lifecycle :as life-cycle]
    [joyride.lm :as lm]
@@ -102,6 +103,10 @@
     (doseq [lm-disposable (lm/register-tools! extension-context)]
       (swap! db/!app-db update :disposables conj lm-disposable)
       (.push (.-subscriptions ^js extension-context) lm-disposable))
+    ;; Register flare sidebar provider
+    (let [flare-disposable (flare-sidebar/register-flare-provider! extension-context)]
+      (swap! db/!app-db update :disposables conj flare-disposable)
+      (.push (.-subscriptions ^js extension-context) flare-disposable))
     (when context (-> (content-utils/maybe-create-user-project+)
                       (p/catch
                        (fn [e]