+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Joyride
 
 ## [Unreleased]
 
+- [Add quick way to create webview content using hiccup (flares)](https://github.com/BetterThanTomorrow/joyride/pull/237)
+
 ## [0.0.63] - 2025-09-08
 
 - [AI: Establish human language names for the tools](https://github.com/BetterThanTomorrow/joyride/pull/234)